reorganize files to packages
