+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".NewCycleActivity" />
-        <activity android:name=".NextExpectedCycleActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.NewCycleActivity" />
+        <activity android:name=".activities.NextExpectedCycleActivity" />
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
