+++ app/src/main/java/com/example/mge_menstracker/calendars/DayViewContainer.kt
@@ -1,7 +1,8 @@
-package com.example.mge_menstracker
+package com.example.mge_menstracker.calendars
 
 import android.view.View
 import android.widget.TextView
+import com.example.mge_menstracker.R
 import com.kizitonwose.calendarview.ui.ViewContainer
 
 class DayViewContainer(view: View) : ViewContainer(view) {
